detection/requirements.txt
ultralytics
kafka-python
opencv-python